:art: #20 - Create change user picture component.
